feat: connects logger to debug flag
